+++ kit/preprocessors/index.js
@@ -1,0 +1,5 @@
+export { frameworkcontentPreprocess } from "./frameworkcontent.js";
+export { docstringPreprocess } from "./docstring.js";
+export { inferenceSnippetPreprocess } from "./inferenceSnippet.js";
+export { tokenizersLangPreprocess } from "./tokenizersLang.js";
+export { mdsvexPreprocess } from "./mdsvex/index.js";
